requirements.txt
torch
torchvision
numpy
Pillow
scikit-learn
onnx
onnxruntime
gradio
Flask
